Version 0.04 - August 2012
1. Minor improvements and bug fixes after lab testing.

2. Allow tools to be piped together into blocks so that not all intermediate
files need to be produced.  For example, freebayes is usually run on a region,
so bamtools is used to merge together bam files in a region and output to a
stream.  The stream is then usually passed through several tools including
bamleftalign and ogap before being passed to freebayes.  There is no need to
create all the intermediate bam files (this would just waste time and space as
the files would be deleted afterwards anyway).