Version 0.24 - September 2012
1. Modified the stdout and stderr from each task in a pipeline so that
they can be written to a single file.  This is the default behaviour.
The pipeline command line argument --task-stdout (-ts) can be set to True
if an individual stdout and stderr from each task in the pipeline is
required.

2. Introduced instances into the pipelines.  Named instances can be included
at the end of a pipeline configuration file and contain different values for
any of the arguments set in the configuration file.