Version 0.41 - October 2012
1. Modified configuration files.

2. If a command line argument is a stub and is also a dependent file for any tasks,
and the value hasn't been set, the makefile still included these as dependencies.
These would just appear as the output extensions.  For example if the jump database is
not used in an alignment task, the dependency list would include '_keys.jmp', '_meta.jmp'
and '_positions.jmp' which would not exist causing the script to fail.  This has been fixed.