Branch: internal loop, commit 10, March 2013
1. Modified the generateCommand subroutine to handle the internal loop.

2. Included a subroutine (checkMakefileID) to ensure that and parameters including
the MAKEFILE_ID variable include the iteration number as well.  This ensure that
the values remain unique even if tasks in the internal loop are all using the
MAKEFILE_ID to define names.  For example, if a temp directory is used and is called
'$(MAKEFILE_ID)_tmp', this will ensure that if multiple makefiles are generated, each
makefile will have a different name for the temp directory.  If the task using the
temp directory appears inside an internal loop, however, there will be several parallel
instances of the task all using the same temp directory, which needs to be avoided.
By modifying the command to also include the iteration number, this value remains
unique for all makefiles and tasks contained in internal loops.