Version 0.56 - April 2013
1. Allowed the linkage for an argument to link to multiple targets.  For example,
TANGRAM uses two reference fastas for generating its own reference file.  The next
task in a pipline merges these two files together to generate a single reference
file for alignment.  The input to the task is both of the files and so the linkage
lists two files to link to.