Version 0.69 - May 2013
1. Added the 'additional dependencies' section to the pipeline configurtaion
files.  This allows additional task dependencies to be defined.  In particular,
some tools (e.g. TANGRAM_DETECT) require the input BAM files to be indexed, but
the indexed file is not included as a command line argument (its existence is
assumed).  This would mean that the index file is not included as a dependency.
When the makefile generated by GKNO is executed, TANGRAM can be executed prior
to the BAM file indexing step as make is unaware that the output of the indexing
is required.