Version 0.84 - June 2013
1. Modified the construction of tool configuration files for tools that are
listed as only outputting to the stream.  If the tool appears as a standalone
task, or at the end of a loop, the tool will write to stdout (and a filename
is defined).

2. The pipeline fastq-bam was added.  This is essentially the same as the
fastq-vcf pipeline, except that the pipeline finishes at the left-aligned
BAM and does not proceed to call variants with FREEBAYES.