Version 0.89 - June 2013
1. Modified the logic when writing command lines to the makefile.  For tools
that are listed as only outputting to stream, but include instructions for
when not outputting to stream, the logic has changed.  Originally, the output
is a redirection to a file, but for some of the tools, this fails.  The reason
for the failure is under investigation with the tool authors, but as a temporary
fix, the redirection first gets piped through cat, solving the problem.

This should be a temporary fix and removed in future versions.