Branch graph, Version 0.29 - November 2013
1. Determined makefile structure and started including routines to write
out data to makefiles. An improvement over the original incarnation of
gkno is that a pipeline can be run in multiple runs mode and gkno will
determine if it needs to be partitioned into phases. For example, phase 1
might consist of building reference files, then phase 2 is multiple
alignment tasks that can be run in parallel. Finally, phase 3 may
collecyt all of the information from phase 2 and perform variant calling
on all of the alignments.