Version 0.112 - January 2014
1. When a task produces multiple outputs, the output list consists of one
of the files and then there is an additional rule to ensure that the
other files exist. If this additional rule is executed first, there is a
recursive use of make and consequently the messages stating that make is
entering and leaving a directory are printed out. This has been made silent.