Version 0.137 - February 2014
1. A host of updates to support an alternate type of filename stub. For one of the
tools, the output file is listed as x.y, but the files produced are x_1.y and x_2.y.
In order to correctly link these files through the pipeline, the argument is treated
as a filename stub with an extension. A number of updates were required to get this
to work as desired.

2. Updated configuration files and added a configuration file for simulation.

3. Updated configurationClass.