Version 0.175 - April 2014
1. Added the primary output to the list of dependencies when writing a rule for additional
outputs (when using a file stub argument). If gkno is run with multiple threads, a task that
produces multiple outputs could be being executed multiple times in parallel. By adding the
output of the main rule to the list of dependencies for the 'additional outputs' rule, the
'additional outputs' rule won't be checked until the task has had a chance to run.