Version 1.30.6 - June 2014
1. Bug fix.

2. Updated gkno configuration file to allow the --makefile-id (-mi) argument.
This allows the user to add an id to the makefile names. Running gkno will create
makefiles based on the tool/pipeline name. Rerunning gkno will overwrite the
makfile. This is good if running on the same data, since it will rerun the same
pipeline, but only required steps (no previously completed steps are rerun). If
the same too/pipeline is to be run on different data, overwriting the makefile
is not desirable, so the user can add an id to generate new makefiles.