Version 1.30.12 - July 2014
1. Bug fix for pipelines with a greedy input argument with multiple values
and a non-greedy argument also with multiple arguments. For example a pipeline
where a task takes multiple input file iterations (from a previous phase) and
uses all values in the executable command line, but there is also another
option paramter with multiple values, meaning that multiple output files are
produced for each value associated with this non-greedy argument.