Version 1.30.18 - July 2014
1. Allowed configuration files to include instructions on how to print the supplied value
on the command line. In particular, this supports providing compressed fastq files to
Jellyfish. If the provided file is a fastq file, the command line just includes the fastq
file. If the file is compressed (.gz), then the file is written to the command line as an
instruction including decompression instructions.

2. R script added for plotting k-mer coverages for trio analysis.

3. Added scripts for copying or renaming files.