Version branch.version2 0.23 - December 2014
1. Finished importing arguments from a tool (not handled from multiple tools or pipelines).

2. Added some more error handling.

3. Updated help messages.

Note that with these updates, the concept of distinct tool and pipeline operation of gkno has
been eliminated. Only pipelines exist. To run an individual tool, a pipeline configuration file
is still required. This pipeline can consist of a single tool and since it is now possible to
import the arguments from the tool configuration file, this configuration file will be very small.