Version branch.version2 0.29 - December 2014
1. A whole set of updates for handling filename construction, in particular ensuring that there are
as many input index file nodes as there are file nodes for the files requiring indexes. For example
a task may have a number of unique nodes specified that all use the same input argument. If the
file associated with this argument have indexes, an index node is created for each of these nodes.

2. Other data QC and plotting routines updated.