Version branch.version2 0.37 - January 2015
1. With the graph constructed, the number of values supplied to all arguments are
checked and the number of times each task is to be executed is determined. For
example, multiple input files could be provided, but if the task is greedy, the
task is executed once using all values. However, multiple values could have been
supplied to an option, forcing multiple executions.

Having determined the number of executions required for each task, and no conflicts
are found, the output files are constructed from the provided instructions. If a
task is being executed multiple times for different values of an argument, that
argument and its value are added to the filename construction instructions.