Version branch.version2 0.57 - February 2015
1. Update to handling multiple input data sets. A pipeline can be run with multiple
input files given to an argument whose argument does not accept multiple input files.
The task will now be split into multiple subphases, so that the pipeline can be run
multiple times for multiple input files without forcing any special requirements on
the user. This is not yet complete and requires testing for, in particular, cases
where multiple options are also supplied.