Version branch.version2 0.66 - February 2015
1. Improved handling of configuration files. If a directory is set for where the
configuration files reside, gkno will check for files in this directory prior to
the default gkno directories. Before processing the files, gkno checks that the
configuration files are for tools or pipelines as required.

2. Configuration files require the 'configuration type' in the top level fields.
The accompanying value must be either 'tool' or 'pipeline'. These are now checked
prior to file processing to ensure the files are of the expected type.