Version branch.version2 0.93 - March 2015
1. Allowed shared nodes in the configuration file to specify a text string to
add to output files constructed for the node. This is useful in cases where a
tasks that start with the same files and perform the same tasks with different
parameters. The filenames produced by the tasks can end up as the same which
will cause problems. If this happens, gkno terminates with an error. If the
node has additional text identified, however, the filenames will be different
and the problem is avoided.