Version branch.version2 0.108 - April 2015
1. A number of fixes and improvements.

2. Random strings in filenames are handled differently. The random strings are
included in all filenames until the end. There are instructions at the end of the
makefile to rename the files, removing the random string.

3. Configuration files updated or removed.

4. The -k flag is added to the makefile execution command. This forces the makefile
to continue all jobs it can after a failure. This will ensure that completed files
have the random string removed, since subsequent runs of gkno will produce different
strings.