Version 2.3.1 - April 2015
1. Updated command line parsing to allow multiple values to be given to a single
argument. For example:

gkno index-bam --in 1.bam 2.bam

is now valid and the two bam files will be associated with the argument --in. The
primary motivation for this change is to allow the use of the wildcard on the command
line. For example:

gkno index-bam --in *bam

will index all bam files in the directory. Additional logic needs to be included to
handle pipelines where multiple arguments are given a wildcard and the arguments are
linked. For example:

gkno bwa -q *_1.fq -q2 *_2.fq

is a valid command line and will result in valid makefiles. In this case, the ordering
of the values is important as each name_1.fq id associated with name_2.fq.