Version 2.9.2 - May 2015
1. Finished adding the glia pipeline.

2. Added a routine to check streams. In the case where a task that runs a pipeline accepts
or outputs to a stream, the first (last) task in that nested pipeline needed to be marked
with stream information. In addition, if a nested pipeline accepts a stream, the first task
is forced to not be greedy. If the task is part of a stream, it can only accept the stream
and not multiple arguments. The nested pipeline if run in isolation could be greedy, so
the isGreedy flag is forced to False.