Version 2.15.0 - June 2015
1. Updated pipeline configuration files to allow the 'greedy task' attribute. It was previously
assumed that if a greedy argument was supplied, the task is greedy, but this is not necessarily
the case. An argument can be greedy, but the input that drives the output files may itself not
be greedy. In that case, the task is not greedy, even though an argument is.

2. Updated configuration files to be consistent with the changes made for 1.