Version 2.20.1 - July 2015
1. Updated handling of files in parameter sets to ensure that $(RESOURCES) or $(PWD) is replaced
with a real path. This is required for lists included in parameter sets.

2. Updated handling of lists and propogation of input files. If a task has a single input with
multiple options and multiple output files specified, the number of input files must match the
number of outputs. Single input files are propogated to create multiple subphases, but the
input file is the same for all subphases.

3. Added bedtool-coverage pipeline.

4. Added find-cds pipeline to interrogate a gff file and output all CDS region coordinates, gene ids,
transcript ids etc.

5. Added the get-region-coverage pipeline to use the outputs of find-cds to generate a file containing
the mean, median, standard deviation, min, max coverage and length of each supplied region.

6. Added scripts to power the above pipelines.